requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
anthropic>=0.39.0
mcp>=1.0.0
python-dotenv>=1.0.0
